pyproject.toml
[project]
name = "saved-ai-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.16.0",
    "backoff>=2.2.1",
    "langchain-openai>=0.2.14",
    "langchain>=0.3.13",
    "openai>=1.58.1",
    "pinecone>=5.4.2",
    "python-dotenv>=1.0.1",
    "tortoise-orm>=0.23.0",
    "langchain-community>=0.3.13",
    "langchain-pinecone>=0.0.1",
    "aiocsv>=1.3.2",
    "apscheduler>=3.11.0",
    "apscheduler-di>=0.1.0",
    "redis>=5.2.1",
    "duckduckgo-search>=7.1.1",
    "tavily-python>=0.5.0",
]
